+++ Civi/DataProcessor/FilterHandler/AbstractFilterHandler.php
@@ -339,6 +339,16 @@ abstract class AbstractFilterHandler {
             $isFilterSet = TRUE;
           }
           break;
+        case 'not null':
+          if (empty($submittedValues['value'])) {
+            $filterParams = [
+              'op' => 'IS NOT NULL',
+              'value' => '',
+            ];
+            $this->setFilter($filterParams);
+            $isFilterSet = TRUE;
+          }
+          break;
         case 'bw':
           if (isset($submittedValues['min']) && $submittedValues['min'] && isset($submittedValues['max']) && $submittedValues['max']) {
             $filterParams = [
@@ -564,6 +574,7 @@ abstract class AbstractFilterHandler {
         'IN' => E::ts('Is one of'),
         'NOT IN' => E::ts('Is not one of'),
         'null' => E::ts('Is empty'),
+        'not null' => E::ts('Is not empty'),
       );
     }
     $types = \CRM_Utils_Type::getValidTypes();
@@ -585,6 +596,7 @@ abstract class AbstractFilterHandler {
           '>' => E::ts('Is greater than'),
           '!=' => E::ts('Is not equal to'),
           'null' => E::ts('Is empty'),
+          'not null' => E::ts('Is not empty'),
           'bw' => E::ts('Is between'),
           'nbw' => E::ts('Is not between'),
         );
@@ -598,6 +610,7 @@ abstract class AbstractFilterHandler {
       'ew' => E::ts('Ends with'),
       'nhas' => E::ts('Does not contain'),
       'null' => E::ts('Is empty'),
+      'not null' => E::ts('Is not empty'),
     );
   }